Apply spacing when adding declarations
Keep source-file batching generic and default declaration trivia in addDeclaration without changing imports or exports.


Copilot-Session: 731b95d5-d943-4b6f-97b7-b923e482069a
